Cleanup types
