+++ apps/web/registry/data-table-filter-v2/lib/array.ts
@@ -21,3 +21,7 @@ export function addUniq<T>(arr: T[], values: T[]): T[] {
 export function removeUniq<T>(arr: T[], values: T[]): T[] {
   return arr.filter((v) => !values.includes(v))
 }
+
+export function isAnyOf<T>(value: T, values: T[]): boolean {
+  return values.includes(value)
+}
